Change order of step by steps in parenting and guardianship topic
Currently 'Get a divorce' is the first step by step listed in this topic. I think it makes more sense to put it below the other two step by steps (Get childcare and Get Tax-Free Childcare), which are more directly relevant to this topic and will probably be useful to more users.